sets the account birthday on block connect/disconnect
sets the account birthday to the earliest transaction timestamp when a block is
connected or disconnected if the following are all true:
- the account was involved in a transaction on the block
- the account birthday is null or the block timestamp is earlier than the
  account birthday

defines 'getEarliestTransactionTimestamp' methods for reading the earliest
timestamp from the wallet database using the timestampToTransactionHash index.

defines wallet methods for updating the account createdAt field. updateAccount
updates the account record in the wallet database and updates the entry in the
wallet's cache.