removes wallet:repair command
we added the 'wallet:repair' command at the end of the phase 2 testnet to
repair known wallet data issues without the need for 'wallet:rescan --reset'
which had a very long runtime on the phase 2 chain.

we've made significant changes to the wallet since the end of phase 2. some of
the issues that we ran into during phase 2 (e.g., balances that did not match
on-chain transactions) no longer apply. we've also added indexes that
'wallet:repair' doesn't address when modifying related datastores. for example,
'wallet:repair' will update decrypted notes to mark them as unspent, but it does
not update the 'unspentNoteHashes' index.

wallet data issues can be resolved by running 'wallet:rescan --reset'. while the
size of the chain database is small this command will not be prohibitively slow
to run.